Move parameter validation to model_functions.py, improve display in CI
